+++ tests/testthat/test-read.R
@@ -31,6 +31,14 @@ test_that("Reading example H5AD works", {
     validateH5ADSCE(sce, names, missing)
 })
 
+test_that("Reading H5AD works with version 0.10.6", {
+    sce <- readH5AD(file, version = "0.10.6")
+    expect_s4_class(sce, "SingleCellExperiment")
+
+    expect_identical(assayNames(sce), "X")
+    expect_identical(colnames(colData(sce)), "cell_type")
+})
+
 test_that("Reading H5AD works with version 0.10.2", {
     sce <- readH5AD(file, version = "0.10.2")
     expect_s4_class(sce, "SingleCellExperiment")